Former-commit-id: c46d3d1d918cea7557de55c06bcd10f3e5a23e1b
Former-commit-id: ace599a636539a86f3de90c6c58e1ec0f56573d2

+++ src/ngamsPlugIns/ngamsJob_MWA_RTS_Task.py
@@ -38,7 +38,7 @@ Some constraints for scheduler:
 
 Either the job itself or the job wrapper should do the additional tasks:
 3. The image result will be archived back to the NGAS node under a different storage set (volume)
-    3.1 Before being archived, the result must be zipped into a single file, whose id is: job_id_obs_id_corr_id.tar.gz
+    3.1 Before being archived, the result must be zipped into a single file, whose id is: job_id_obs_id_corr_id.tar
 """
 import commands, os, fnmatch
 from optparse import OptionParser
@@ -154,10 +154,17 @@ def main():
     
     # create symbolic links in the work_dir to real file paths 
     fileList = opts.file_list.split(',')
+    validFiles = 0
     for file in fileList:
-        base_name = os.path.basename(file)
-        cmd = 'ln -s %s %s' % (file, base_name)
-        execCmd(cmd) #don't care if this failed
+        if (os.path.exists(file)):
+            validFiles += 1
+            base_name = os.path.basename(file)
+            cmd = 'ln -s %s %s' % (file, base_name)
+            execCmd(cmd) #don't care if this failed
+    
+    if (not validFiles):
+        print('No visibility files are there to be processed')
+        exit(116)
     
     # create the template list file
     if (opts.tpl_pf):
@@ -266,8 +273,8 @@ def main():
         execCmd(cmd, False)
     
     # pack all template directories into a single 'image (gzip)' file   
-    imgFile = '%s__%d__%s.tar.gz' % (opts.job_id, obs_num, gpubox)
-    cmd = 'tar -czf %s' % imgFile
+    imgFile = '%s__%d__%s.tar' % (opts.job_id, obs_num, gpubox)
+    cmd = 'tar -cf %s' % imgFile
     for tpa in temp_alias:
         cmd += ' %s/' % tpa
     execCmd(cmd)
